first stable version

+++ lib/screens/history_screen.dart
@@ -1,3 +1,5 @@
+import 'package:agri_rongeur_mob/utils/constants.dart';
+import 'package:agri_rongeur_mob/utils/dialogs.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -7,15 +9,18 @@ import 'result_screen.dart';
 class HistoryScreen extends StatelessWidget {
   const HistoryScreen({super.key});
 
-  Future<void> _deleteEntry(String docId) async {
-    final uid = FirebaseAuth.instance.currentUser?.uid;
-    if (uid != null) {
-      await FirebaseFirestore.instance
-          .collection('users')
-          .doc(uid)
-          .collection('history')
-          .doc(docId)
-          .delete();
+  Future<void> _deleteEntry(String docId, BuildContext context) async {
+    final confirmed = await MyDialog.confirmAndDelete(context, docId);
+    if (confirmed ?? false) {
+      final uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid != null) {
+        await FirebaseFirestore.instance
+            .collection('users')
+            .doc(uid)
+            .collection('history')
+            .doc(docId)
+            .delete();
+      }
     }
   }
 
@@ -35,8 +40,29 @@ class HistoryScreen extends StatelessWidget {
       body: StreamBuilder<QuerySnapshot>(
         stream: historyRef.snapshots(),
         builder: (context, snapshot) {
-          if (snapshot.hasError) return const Text('Erreur');
-          if (!snapshot.hasData) return const CircularProgressIndicator();
+          if (snapshot.hasError) {
+            return const Center(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Icon(Icons.warning, color: Color(0xFFF44336),),
+                  Text('Erreur'),
+                ],
+              ),
+            );
+          }
+          if (!snapshot.hasData) {
+            return const Center(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  CircularProgressIndicator(),
+                ],
+              ),
+            );
+          }
 
           final docs = snapshot.data!.docs;
 
@@ -49,28 +75,26 @@ class HistoryScreen extends StatelessWidget {
             itemBuilder: (context, index) {
               final doc = docs[index];
               final data = doc.data() as Map<String, dynamic>;
-              final result = data['result'] ?? {};
+              final result = data['result'] ?? [];
               final imageUrl = data['image_url'];
-              final name = result['name'] ?? 'Inconnu';
+              // final name = result['name'] ?? 'Inconnu';
               final createdAt = (data['createdAt'] as Timestamp?)?.toDate();
 
               return ListTile(
                 leading: imageUrl != null
-                    ? Image.network(imageUrl,
+                    ? Image.network("${AppConstants.serverBaseUrl}$imageUrl",
                         width: 50, height: 50, fit: BoxFit.cover)
                     : const Icon(Icons.image),
-                title: Text(name),
-                subtitle: createdAt != null
-                    ? Text('${createdAt.toLocal()}'.split('.')[0])
-                    : null,
+                title: Text(createdAt != null ? '${createdAt.toLocal()}'.split('.')[0] : "Date non définie"),
+                subtitle: Text("${result.length} rongeur(s) détecté(s)"),
                 trailing: IconButton(
-                  icon: const Icon(Icons.delete),
-                  onPressed: () => _deleteEntry(doc.id),
+                  icon: const Icon(Icons.delete_forever),
+                  onPressed: () => _deleteEntry(doc.id, context),
                 ),
                 onTap: () => Navigator.push(
                   context,
                   MaterialPageRoute(
-                    builder: (_) => ResultScreen(result: result),
+                    builder: (_) => ResultScreen(result: data),
                   ),
                 ),
               );
